Final state for session 3 delivery - 13/03/24-12:09

+++ course-project/algstudentUO295180/src/algstudent/s3/Fibonacci1.java
@@ -9,18 +9,18 @@ public class Fibonacci1 {
 	  int n= Integer.parseInt(arg[0]);
 	    
 	  System.out.println("The Fibonacci number of order "+n+" is " + fib1(n));
-	  int[]v = new int[60];  
+	  long[]v = new long[60];  
 	  System.out.println("The Fibonacci number of order "+n+" is " + fib2(n,v));
 	  System.out.println("The Fibonacci number of order "+n+" is " + fib3(n));
 	  System.out.println("The Fibonacci number of order "+n+" is " + fib4(n));
 	}   
 	
 	/* Iterative solution with temporal complexity O(n) */
-	public static int fib1(int n) {
-	  int n1 = 0;
-	  int n2 = 1;
+	public static long fib1(int n) {
+	  long n1 = 0;
+	  long n2 = 1;
 	  for (int i = 1; i <= n; i++) {
-	      int s= n1+n2;
+	      long s= n1+n2;
 	      n1 = n2;
 	      n2 = s;
 	  }
@@ -30,7 +30,7 @@ public class Fibonacci1 {
 	/* Second iterative solution with time complexity O(n) 
 	 * using a vector. This algorithm will be viewed later 
 	 * as a very simple case dynamic programming */
-	public static int fib2(int n, int[]v) {
+	public static long fib2(int n, long[]v) {
 	  v[0] = 0;
 	  v[1] = 1;
 	  for (int i=2; i <= n; i++) 
@@ -41,11 +41,11 @@ public class Fibonacci1 {
 	/* First recursive solution: T(n)=T(n-1)+0(1)
 	 * (a=1;b=1;k=0)=>O(n) 
 	 * => O(n) */
-	public static int fib3(int n) {
+	public static long fib3(int n) {
 	    return aux(0, 1, n);
 	}
 	
-	private static int aux(int n1, int n2, int n) {
+	private static long aux(long n1, long n2, int n) {
 	  if (n < 1) return n1;
 	    return aux(n2, n1+n2, n-1);
 	}
@@ -57,7 +57,7 @@ public class Fibonacci1 {
 	 * for n is the sum of times for the preceding sizes 
 	 * (n-1) and (n-2) 
 	 * (a=2;b=1 or b=2;k=0) */
-	public static int fib4(int n) {
+	public static long fib4(long n) {
 	  if (n<=1)
 	     return n;
 	  return fib4(n-1) + fib4(n-2);
